refactor: split manifest, DRY credential extraction, fix test race
- Extract manifest declaration into manifest.go (~180 lines) to keep
  figma.go focused on connector logic and improve maintainability
- Extract requireToken() helper to DRY up credential validation in
  ValidateCredentials, doGet, and doPost
- Fix data race in TestFigmaConnector_RedirectStripsToken: use
  atomic.Bool instead of plain string for cross-goroutine state
